# perf(parser): cache regex predicates with rc using thread_local