@nowNick 2024-10-09
Closing this PR in favor of https://github.com/Kong/atc-router/pull/251 as that one is the preferred solution.